refactor: 테스트 코드 및 컴포넌트 분리
