+++ app/src/main/java/nextstep/signup/view/PasswordConfirmTextField.kt
@@ -15,9 +15,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.signup.R
-import nextstep.signup.TextFieldState
-import nextstep.signup.TextFieldType
-import nextstep.signup.setMessage
+import nextstep.signup.model.TextFieldState
+import nextstep.signup.model.TextFieldType
+import nextstep.signup.model.setMessage
 import nextstep.signup.view.ui.theme.Red50
 import nextstep.signup.view.ui.theme.SignupTheme
 
@@ -39,6 +39,15 @@ fun PasswordConfirmTextField(
         onTextChange(value)
     }
 
+    PasswordConfirmContent(textState, onValueChange, validationState)
+}
+
+@Composable
+fun PasswordConfirmContent(
+    textState: String,
+    onValueChange: (String) -> Unit,
+    validationState: TextFieldState
+) {
     TextField(
         modifier = Modifier.fillMaxWidth(),
         value = textState,
