+++ app/src/main/java/nextstep/signup/view/UsernameTextField.kt
@@ -13,11 +13,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import nextstep.signup.InputValidator.isValid
+import nextstep.signup.model.InputValidator.isValid
 import nextstep.signup.R
-import nextstep.signup.TextFieldState
-import nextstep.signup.TextFieldType
-import nextstep.signup.setMessage
+import nextstep.signup.model.TextFieldState
+import nextstep.signup.model.TextFieldType
+import nextstep.signup.model.setMessage
 import nextstep.signup.view.ui.theme.Red50
 import nextstep.signup.view.ui.theme.SignupTheme
 
@@ -31,6 +31,15 @@ fun UsernameTextField() {
         validationState = isValid(value, TextFieldType.Username)
     }
 
+    UsernameContent(textState, onValueChange, validationState)
+}
+
+@Composable
+fun UsernameContent(
+    textState: String,
+    onValueChange: (String) -> Unit,
+    validationState: TextFieldState
+) {
     TextField(
         modifier = Modifier.fillMaxWidth(),
         value = textState,
